refactor: css processing, remove dependency on old compiled css files
