+++ gatsby/package.json
@@ -10,8 +10,10 @@
   "dependencies": {
     "@mdx-js/mdx": "^1.0.18",
     "@mdx-js/react": "^1.0.16",
+    "bootstrap": "^3.4.1",
     "date-fns": "^1.30.1",
     "dotenv": "^6.2.0",
+    "font-awesome": "^4.7.0",
     "gatsby": "^2.1.31",
     "gatsby-image": "^2.0.33",
     "gatsby-mdx": "^0.6.3",
@@ -34,6 +36,8 @@
     "gatsby-transformer-remark": "^2.3.2",
     "gatsby-transformer-sharp": "^2.1.17",
     "gatsby-transformer-yaml": "^2.1.10",
+    "highlight.js": "^9.15.8",
+    "node-font-awesome": "^1.0.2",
     "prismjs": "^1.15.0",
     "react": "^16.8.4",
     "react-dom": "^16.8.4",
