+++ gatsby/src/layout/footer/index.js
@@ -1,6 +1,6 @@
 import { graphql, StaticQuery } from "gatsby"
 import React from "react"
-import MarketoForm from "./marketoForm"
+import MarketoForm from "../../components/marketoForm"
 
 const Footer = ({ data }) => {
   const pantheonLogo = data.allFile.edges.find(
