+++ gatsby/src/layout/layout/index.js
@@ -1,6 +1,7 @@
 import React from "react"
-import Header from "./header"
-import Footer from "./footer"
+import Header from "../header"
+import Footer from "../footer"
+import './style.css'
 
 const Layout = ({ children }) => {
   return (
