+++ gatsby/src/layout/layout/style.css
@@ -1,0 +1,9 @@
+@import '../../../node_modules/highlight.js/styles/tomorrow-night-bright.css';
+@import '../../../node_modules/font-awesome/css/font-awesome.min.css';
+@import '../../../node_modules/bootstrap/dist/css/bootstrap.min.css';
+@import '../../styles/custom-glyphicons.css';
+@import '../../../../source/docs/assets/fonts/pan-docs-icons/css/pan-docs.css';
+@import '../../../../source/docs/assets/fonts/MyFontsWebfontsKit/MyFontsWebfontsKit.css';
+@import '../../styles/main.css';
+
+
