+++ gatsby/src/pages/404.js
@@ -1,8 +1,8 @@
 import React from "react"
 // import { graphql } from "gatsby"
 
-// import Layout from "../components/layout"
-// import SEO from "../components/seo"
+// import Layout from "../layout/layout"
+// import SEO from "../layout/seo"
 
 class NotFoundPage extends React.Component {
   render() {
