+++ gatsby/src/templates/contributor.js
@@ -1,6 +1,6 @@
 import React from "react"
 import { graphql, Link } from "gatsby"
-import Layout from "../components/layout"
+import Layout from "../layout/layout"
 import ContributorLink from "../components/contributorLink"
 
 const links = [
