+++ gatsby/src/templates/doc.js
@@ -3,7 +3,7 @@ import { graphql } from "gatsby"
 import MDXRenderer from "gatsby-mdx/mdx-renderer"
 import { MDXProvider } from "@mdx-js/react"
 
-import Layout from "../components/layout"
+import Layout from "../layout/layout"
 import Callout from "../components/callout"
 import Alert from "../components/alert"
 import Accordion from "../components/accordion"
@@ -20,7 +20,7 @@ import Slack from "../components/slack"
 import GetFeedback from "../components/getFeedback"
 import Card from "../components/card"
 import CardGroup from "../components/cardGroup"
-import SEO from "../components/seo"
+import SEO from "../layout/seo"
 import DefList from "../components/defList.js"
 import Definition from "../components/definition.js"
 import Enablement from "../components/enablement.js"
