+++ gatsby/src/templates/terminuspage.js
@@ -3,7 +3,7 @@ import { graphql } from "gatsby"
 import MDXRenderer from "gatsby-mdx/mdx-renderer"
 import { MDXProvider } from "@mdx-js/react"
 
-import Layout from "../components/layout"
+import Layout from "../layout/layout"
 import Callout from "../components/callout"
 import Alert from "../components/alert"
 import Accordion from "../components/accordion"
@@ -21,7 +21,7 @@ import Card from "../components/card"
 import CardGroup from "../components/cardGroup"
 import Navbar from "../components/navbar"
 import NavButtons from "../components/navButtons"
-import SEO from "../components/seo"
+import SEO from "../layout/seo"
 import Releases from "../components/releases"
 import TerminusVersion from "../components/terminusVersion"
 import Commands from "../components/commands"
